fix: bag and add filters

+++ db_control.py
@@ -11,7 +11,7 @@ class Games(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(60), index=True, unique=True)
     dev = db.Column(db.String(60), index=True, unique=True)
-    price = db.Column(db.Float, index=True)
+    price = db.Column(db.Integer, index=True)
 
 
 class User(db.Model, UserMixin):
